ref(dashboards): Replace renderExtraActions function prop with component
Replace the render function prop with a boolean `showDashboardLinkButton`
and a colocated `LinkToDashboardAction` component. This avoids recreating
the render function on every parent render and simplifies the API.
